backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
pydantic==2.5.0
pydantic-settings==2.1.0
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
better-exceptions==0.3.3
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
bcrypt==4.0.1
google-generativeai==0.8.4
aiosqlite==0.19.0